requirements.txt
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
tqdm==4.66.1
fake-useragent==1.4.0